Build the transfer source and destination info displayed at the redeem view on the routes instead of the SendFrom/SendTo components
