+++ wormhole-connect/src/views/Redeem/SendTo.tsx
@@ -1,205 +1,27 @@
+import CircularProgress from '@mui/material/CircularProgress';
+import { Context } from '@wormhole-foundation/wormhole-connect-sdk';
+import { utils } from 'ethers';
 import React, { useCallback, useEffect, useState } from 'react';
-import { useSelector, useDispatch } from 'react-redux';
-import { BigNumber, utils } from 'ethers';
-import {
-  ChainName,
-  Context,
-  WormholeContext,
-  SuiContext,
-  AptosContext,
-} from '@wormhole-foundation/wormhole-connect-sdk';
+import { useDispatch, useSelector } from 'react-redux';
+import AlertBanner from '../../components/AlertBanner';
+import Button from '../../components/Button';
+import InputContainer from '../../components/InputContainer';
+import { RenderRows } from '../../components/RenderRows';
+import Spacer from '../../components/Spacer';
+import { CHAINS } from '../../config';
 import { RootState } from '../../store';
 import { setRedeemTx, setTransferComplete } from '../../store/redeem';
 import { Route } from '../../store/transferInput';
+import { displayAddress } from '../../utils';
+import { fetchRedeemTx } from '../../utils/events';
+import Operator, { TransferDisplayData } from '../../utils/routes';
 import {
-  MAX_DECIMALS,
-  displayAddress,
-  fromNormalizedDecimals,
-  getTokenDecimals,
-  getWrappedTokenId,
-  toNormalizedDecimals,
-} from '../../utils';
-import {
+  TransferWallet,
   registerWalletSigner,
   switchNetwork,
-  TransferWallet,
 } from '../../utils/wallet';
-import { toDecimals } from '../../utils/balance';
-import { fetchRedeemTx, fetchSwapEvent } from '../../utils/events';
-import { wh, calculateNativeTokenAmt, toChainId } from '../../utils/sdk';
-import { CHAINS, TOKENS, GAS_ESTIMATES } from '../../config';
 import WalletsModal from '../WalletModal';
-
 import Header from './Header';
-import AlertBanner from '../../components/AlertBanner';
-import CircularProgress from '@mui/material/CircularProgress';
-// import Confirmations from './Confirmations';
-import Button from '../../components/Button';
-import Spacer from '../../components/Spacer';
-import { RenderRows, RowsData } from '../../components/RenderRows';
-import InputContainer from '../../components/InputContainer';
-import { Types } from 'aptos';
-import { getTotalGasUsed } from '@mysten/sui.js';
-import { estimateClaimGasFees } from '../../utils/gasEstimates';
-import Operator from '../../utils/routes';
-
-const calculateGas = async (chain: ChainName, receiveTx?: string) => {
-  const { gasToken } = CHAINS[chain]!;
-  const token = TOKENS[gasToken];
-  const decimals = getTokenDecimals(toChainId(chain), token.tokenId);
-
-  if (chain === 'solana') {
-    return toDecimals(
-      BigNumber.from(GAS_ESTIMATES.solana!.claim),
-      decimals,
-      MAX_DECIMALS,
-    );
-  }
-  if (receiveTx) {
-    if (chain === 'aptos') {
-      const aptosClient = (
-        wh.getContext('aptos') as AptosContext<WormholeContext>
-      ).aptosClient;
-      const txn = await aptosClient.getTransactionByHash(receiveTx);
-      if (txn.type === 'user_transaction') {
-        const userTxn = txn as Types.UserTransaction;
-        const gasFee = BigNumber.from(userTxn.gas_used).mul(
-          userTxn.gas_unit_price,
-        );
-        return toDecimals(gasFee || 0, decimals, MAX_DECIMALS);
-      }
-    } else if (chain === 'sui') {
-      const provider = (wh.getContext('sui') as SuiContext<WormholeContext>)
-        .provider;
-      const txBlock = await provider.getTransactionBlock({
-        digest: receiveTx,
-        options: { showEvents: true, showEffects: true, showInput: true },
-      });
-      const gasFee = BigNumber.from(getTotalGasUsed(txBlock) || 0);
-      return toDecimals(gasFee, decimals, MAX_DECIMALS);
-    } else {
-      const provider = wh.mustGetProvider(chain);
-      const receipt = await provider.getTransactionReceipt(receiveTx);
-      const { gasUsed, effectiveGasPrice } = receipt;
-      if (!gasUsed || !effectiveGasPrice) return;
-      const gasFee = gasUsed.mul(effectiveGasPrice);
-      return toDecimals(gasFee, decimals, MAX_DECIMALS);
-    }
-  }
-  return await estimateClaimGasFees(chain);
-};
-
-const getManualRows = async (
-  txData: any,
-  receiveTx?: string,
-): Promise<RowsData> => {
-  const token = TOKENS[txData.tokenKey];
-  const { gasToken } = CHAINS[txData.toChain]!;
-
-  // get gas used (if complete) or gas estimate if not
-  const gas = await calculateGas(txData.toChain, receiveTx);
-
-  const formattedAmt = toNormalizedDecimals(
-    txData.amount,
-    txData.tokenDecimals,
-    MAX_DECIMALS,
-  );
-  return [
-    {
-      title: 'Amount',
-      value: `${formattedAmt} ${token.symbol}`,
-    },
-    {
-      title: receiveTx ? 'Gas fee' : 'Gas estimate',
-      value: gas ? `${gas} ${gasToken}` : '—',
-    },
-  ];
-};
-
-const getAutomaticRows = async (
-  txData: any,
-  receiveTx?: string,
-  transferComplete?: boolean,
-): Promise<RowsData> => {
-  const token = TOKENS[txData.tokenKey];
-  const { gasToken } = CHAINS[txData.toChain]!;
-
-  // calculate the amount of native gas received
-  let nativeGasAmt: string | undefined;
-  const nativeGasToken = TOKENS[gasToken];
-  if (receiveTx) {
-    let nativeSwapAmount: any;
-    try {
-      nativeSwapAmount = await fetchSwapEvent(txData);
-    } catch (e) {
-      console.error(`could not fetch swap event:\n${e}`);
-    }
-    if (nativeSwapAmount) {
-      const decimals = getTokenDecimals(
-        wh.toChainId(txData.toChain),
-        nativeGasToken.tokenId,
-      );
-      nativeGasAmt = toDecimals(nativeSwapAmount, decimals, MAX_DECIMALS);
-    }
-  } else if (!transferComplete) {
-    // get the decimals on the target chain
-    const destinationTokenDecimals = getTokenDecimals(
-      wh.toChainId(txData.toChain),
-      txData.tokenId,
-    );
-    const amount = await calculateNativeTokenAmt(
-      txData.toChain,
-      txData.tokenId,
-      fromNormalizedDecimals(
-        txData.toNativeTokenAmount,
-        destinationTokenDecimals,
-      ),
-      txData.recipient,
-    );
-    // get the decimals on the target chain
-    const nativeGasTokenDecimals = getTokenDecimals(
-      wh.toChainId(txData.toChain),
-      getWrappedTokenId(nativeGasToken),
-    );
-    nativeGasAmt = toDecimals(
-      amount.toString(),
-      // nativeGasToken.decimals,
-      nativeGasTokenDecimals,
-      MAX_DECIMALS,
-    );
-  }
-
-  const receiveAmt = BigNumber.from(txData.amount)
-    .sub(BigNumber.from(txData.relayerFee))
-    .sub(BigNumber.from(txData.toNativeTokenAmount || 0));
-  const formattedAmt = toNormalizedDecimals(
-    receiveAmt,
-    txData.tokenDecimals,
-    MAX_DECIMALS,
-  );
-
-  return [
-    {
-      title: 'Amount',
-      value: `${formattedAmt} ${token.symbol}`,
-    },
-    {
-      title: 'Native gas token',
-      value: nativeGasAmt ? `${nativeGasAmt} ${gasToken}` : '—',
-    },
-  ];
-};
-
-const getRows = async (
-  txData: any,
-  receiveTx?: string,
-  transferComplete?: boolean,
-): Promise<RowsData> => {
-  if (txData.payloadID === Route.BRIDGE) {
-    return await getManualRows(txData, receiveTx);
-  }
-  return await getAutomaticRows(txData, receiveTx, transferComplete);
-};
 
 function SendTo() {
   const dispatch = useDispatch();
@@ -227,7 +49,7 @@ function SendTo() {
 
   const [inProgress, setInProgress] = useState(false);
   const [isConnected, setIsConnected] = useState(checkConnection());
-  const [rows, setRows] = useState([] as RowsData);
+  const [rows, setRows] = useState([] as TransferDisplayData);
   const [openWalletModal, setWalletModal] = useState(false);
 
   // get the redeem tx, for automatic transfers only
@@ -245,17 +67,21 @@ function SendTo() {
   useEffect(() => {
     if (!txData) return;
     const populate = async () => {
-      let tx: string | undefined;
+      let receiveTx: string | undefined;
       try {
-        tx = await getRedeemTx();
+        receiveTx = await getRedeemTx();
       } catch (e) {
         console.error(`could not fetch redeem event:\n${e}`);
       }
-      const rows = await getRows(txData, tx, transferComplete);
+      const rows = await new Operator().getTransferDestInfo(routeType, {
+        txData,
+        receiveTx,
+        transferComplete,
+      });
       setRows(rows);
     };
     populate();
-  }, [transferComplete, getRedeemTx, txData]);
+  }, [transferComplete, getRedeemTx, txData, routeType]);
 
   useEffect(() => {
     setIsConnected(checkConnection());
